+++ output/models/ms_data/datatypes/double_xsd/double.py
@@ -1,5 +1,4 @@
 from dataclasses import dataclass, field
-from decimal import Decimal
 from typing import Optional
 
 
@@ -8,7 +7,7 @@ class ComplexfooType:
     class Meta:
         name = "complexfooType"
 
-    comp_foo: Optional[Decimal] = field(
+    comp_foo: Optional[float] = field(
         default=None,
         metadata={
             "type": "Element",
@@ -23,7 +22,7 @@ class SimpleTest:
     class Meta:
         name = "simpleTest"
 
-    value: Optional[Decimal] = field(
+    value: Optional[float] = field(
         default=None,
         metadata={
             "required": True,
@@ -50,7 +49,7 @@ class Root:
             "required": True,
         }
     )
-    simple_test: Optional[Decimal] = field(
+    simple_test: Optional[float] = field(
         default=None,
         metadata={
             "name": "simpleTest",
